Fix positioning problem of activating select2
Changed the order of the fired event and the call to activating the control. 
Before thew select2 was activated and after the select2 dropdown was rendered the popover was repositioned when the show event was fired. However the  select2 dropdown position remained unchanged and the end placement was wrong.

With this fix the show event is fired (positioning the popover) before the callback to activating the element (that shows and positions the select2 dropdown)